+++ gluecodium/src/test/resources/smoke/constants/output/cpp/include/smoke/Constants.h
@@ -6,6 +6,7 @@
 #include "gluecodium/ExportGluecodiumCpp.h"
 #include <cstdint>
 #include <string>
+#include <string_view>
 namespace smoke {
 struct _GLUECODIUM_CPP_EXPORT Constants {
     static const bool BOOL_CONSTANT;
@@ -13,7 +14,7 @@ struct _GLUECODIUM_CPP_EXPORT Constants {
     static const uint32_t UINT_CONSTANT;
     static const float FLOAT_CONSTANT;
     static const double DOUBLE_CONSTANT;
-    static const ::std::string STRING_CONSTANT;
+    static constexpr std::string_view STRING_CONSTANT = "Foo bar";
     enum class StateEnum {
         OFF,
         ON