+++ gluecodium/src/test/resources/smoke/constants/output/cpp/src/smoke/Constants.cpp
@@ -10,6 +10,5 @@ const int32_t Constants::INT_CONSTANT = -11;
 const uint32_t Constants::UINT_CONSTANT = 4294967295;
 const float Constants::FLOAT_CONSTANT = 2.71f;
 const double Constants::DOUBLE_CONSTANT = -3.14;
-const ::std::string Constants::STRING_CONSTANT = "Foo bar";
 const ::smoke::Constants::StateEnum Constants::ENUM_CONSTANT = ::smoke::Constants::StateEnum::ON;
 }