+++ gluecodium/src/test/resources/smoke/structs/output/cpp/include/smoke/StructsWithConstants.h
@@ -6,10 +6,11 @@
 #include "gluecodium/ExportGluecodiumCpp.h"
 #include "smoke/RouteUtils.h"
 #include <string>
+#include <string_view>
 namespace smoke {
 struct _GLUECODIUM_CPP_EXPORT StructsWithConstants {
     struct _GLUECODIUM_CPP_EXPORT Route {
-        static const ::std::string DEFAULT_DESCRIPTION;
+        static constexpr std::string_view DEFAULT_DESCRIPTION = "Nonsense";
         static const ::smoke::RouteUtils::RouteType DEFAULT_TYPE;
         ::std::string description;
         ::smoke::RouteUtils::RouteType type;